# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sales Portal MVP API
  version: 1.0.0
  description: >-
    Generated from locked Phase 4 API conventions and Chunk 1–4 contracts, plus the Phase 13
    Invoice Acceptance + Signature + Email addendum
    (docs/API-Contracts-Phase13-Acceptance-Signature-Email.md).
servers:
  - url: /api/v1/{slug}
    variables:
      slug:
        default: aussie-floors-group
        description: Business slug tenant key resolved by the backend.
security:
  - SessionCookieAuth: []
...
paths:
...
  /orders/{orderId}/quote/workspace:
    get:
      tags:
        - Quotes
      summary: Get quote workspace
      operationId: getQuoteWorkspace
      description: >
        Loads the full quote state for the order's Quote tab (Phase 16B): the editable draft, the
        active issued version summary, and the latest accepted version summary (Phase 16F PR1: the
        highest-numbered ACCEPTED version, or null when none; selected independently of the draft
        and of current_issued). The read runs in one read-only REPEATABLE_READ transaction (no row
        locks; never blocks a writer), so all members come from one consistent snapshot and the
        same version never appears as both current_issued and accepted. Cost/GP flags are present
        on this protected view only. Allowed when order_status is LAID (read).
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Quote workspace state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteWorkspaceResponseWrapper'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}/quote/draft:
    put:
      tags:
        - Quotes
      summary: Upsert quote draft
      operationId: upsertQuoteDraft
      description: >
        Creates or replaces the editable quote draft (itemised flag, customer-facing lines and
        adjustments) for the order (Phase 16B). An itemised save full-replaces the draft lines:
        the server recomputes totals, enforces total == sum of lines (auto-inserting a negative
        adjustment line on a direct reduction via final_total_inc_gst) and rejects a total above
        the line sum (QUOTE_TOTAL_EXCEEDS_LINES, 422; itemised saves only). A non-itemised save
        (itemised false, final_total_inc_gst required, lines must be empty) is header-only
        (Phase 16D-B PR2A): it updates mode/totals (inc = final_total_inc_gst; ex = final / 1.10
        HALF_UP 2dp), creates NO synthetic line, and never writes or deletes quote_draft_line
        rows — lines persisted by the last itemised save are retained as dormant rows and are
        returned in the response lines[]. The non-itemised total is never validated against
        retained lines. Below-cost saves are blocked in both modes (QUOTE_BELOW_COST, 422):
        itemised uses the visible line sum, non-itemised the final-total-derived ex total. The
        quote draft is independent of order pricing: a save does NOT update the order sale-price
        override or any sales_order header financials, and quote lines never mutate the order's
        product/charge lines. Blocked when order_status is LAID (ORDER_LOCKED, 422).
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteDraftUpsertRequest'
      responses:
        '200':
          description: Quote draft saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteDraftResponseWrapper'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}/quote/preview-pdf:
    post:
      tags:
        - Quotes
      summary: Preview quote draft PDF
      operationId: previewQuotePdf
      description: >
        Generates an on-demand draft quote PDF (Aire Compact style, title QUOTATION, including a display-only 40% deposit line and printable declaration/customer acceptance area) for preview.
        Not stored. Empty JSON object body. Below-cost does not block a preview (only
        save/send/accept block). Allowed when order_status is LAID (read-only render).
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyObjectRequest'
      responses:
        '200':
          description: Draft quote PDF streamed
          headers:
            Content-Disposition:
              description: Inline preview filename generated by backend.
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}/quote/send-email:
    post:
      tags:
        - Quotes
      summary: Send quote by email
      operationId: sendQuoteEmail
      description: >
        Issues (or resends) the quote and emails the public quote link (Phase 16B; implemented
        16E-A). Link-only since 16F PR1: a short plain-text body carrying the public quote link
        exactly once and NO PDF attachment (16E-A attached the issued PDF). The immutable issued PDF
        is still generated and stored when a new version is issued; the customer opens it from the
        public page (GET /public/quotes/{token}/pdf) and the salesperson via
        GET /orders/{orderId}/quote/pdf?type=issued. The link's token is the credential for the
        public view, viewed, PDF and accept endpoints (acceptPublicQuote implemented 16F PR1; the
        public signing UI is planned 16F PR3). A draft or customer name/billing change since the
        last issue (compared with the V17 issue-time snapshot), or no active ISSUED version at all
        (never sent, cancelled, expired or accepted), creates a new ISSUED version (next
        version_number) that supersedes any previous ISSUED one; otherwise the same version is
        re-sent with a new ACTIVE token (old token REPLACED) and its stored PDF is kept as is
        (never regenerated). Either way the issued version, its stored PDF and the new token are
        persisted before delivery. Sending after an acceptance is allowed (16F decision D9): it
        issues a new version that needs a new signature and never changes the ACCEPTED version. A
        send never writes the order sale price (only acceptPublicQuote does, decision D6b). In the
        current build delivery is recorded in memory only - no real email is sent. Re-checks
        below-cost (QUOTE_BELOW_COST, 422) and requires a valid customer email
        (CUSTOMER_EMAIL_REQUIRED / CUSTOMER_EMAIL_INVALID, 422). Blocked when LAID (ORDER_LOCKED,
        422). On provider failure returns 502 (EMAIL_SEND_FAILED); the issued version/PDF/token
        are kept and the salesperson resends.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyObjectRequest'
      responses:
        '201':
          description: Quote issued and emailed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteSendResponseWrapper'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}/quote/send-sms:
    post:
      tags:
        - Quotes
      summary: Send quote by SMS
      operationId: sendQuoteSms
      description: >
        DORMANT - existing backend support that is excluded from the approved delivery roadmap
        (delivery is email only). The frontend never calls it and no SMS provider exists
        (in-memory recording sender only); the definition is kept, not promised.
        Issues (or resends) the quote and sends an SMS containing the public quote link only (no
        PDF) (Phase 16B). Same issue/resend + token logic as send-email. Requires a valid customer
        mobile (CUSTOMER_MOBILE_REQUIRED / CUSTOMER_MOBILE_INVALID, 422). Blocked when LAID
        (ORDER_LOCKED, 422). On provider failure returns 502 (SMS_SEND_FAILED); the issued
        version/PDF/token are kept and the salesperson resends.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyObjectRequest'
      responses:
        '201':
          description: Quote issued and SMS sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteSendResponseWrapper'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '502':
          $ref: '#/components/responses/BadGateway'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}/quote/cancel:
    post:
      tags:
        - Quotes
      summary: Cancel active issued quote
      operationId: cancelQuote
      description: >
        Cancels the active issued quote (Phase 16B): version becomes CANCELLED and the active
        token becomes CANCELLED (kept for messaging; link dead). Requires an active ISSUED version:
        with none, an order that has an ACCEPTED quote version returns 409
        (QUOTE_ALREADY_ACCEPTED - reachable since 16F PR1 creates ACCEPTED versions), otherwise 422
        (QUOTE_NOT_ISSUED). An ISSUED version sent after an acceptance is cancelled normally; an
        ACCEPTED version is never touched. Never writes the order sale price. Empty JSON object
        body. Allowed when LAID as a narrow exception (kills a public link only; not an order
        mutation).
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyObjectRequest'
      responses:
        '200':
          description: Quote cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteSendResponseWrapper'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /orders/{orderId}/quote/pdf:
    get:
      tags:
        - Quotes
      summary: Download stored quote PDF (salesperson)
      operationId: downloadQuotePdf
      description: >
        Streams a stored quote PDF for the salesperson (Phase 16B): type=issued returns the active
        issued version's stored PDF (inline filename quote-{order_number}-v{n}.pdf); type=accepted
        returns the latest ACCEPTED version's stored signed PDF (implemented 16F PR1 - the
        immutable bytes stored at acceptance, never regenerated; inline filename
        quote-{order_number}-v{n}-signed.pdf). The signed PDF is portal only and is never served on
        the public surface. Missing requested artifact (no active issued version for type=issued;
        no accepted version or no stored signed PDF for type=accepted) returns 404
        (QUOTE_PDF_NOT_FOUND). Allowed when order_status is LAID (read).
      parameters:
        - $ref: '#/components/parameters/OrderId'
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum:
              - issued
              - accepted
      responses:
        '200':
          description: Quote PDF streamed
          headers:
            Content-Disposition:
              description: Inline quote PDF filename generated by backend.
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}/quote/accepted/signature:
    get:
      tags:
        - Quotes
      summary: Download accepted quote signature image
      operationId: downloadAcceptedQuoteSignature
      description: >
        Streams the customer signature image (image/png) stored when the order's latest ACCEPTED
        quote version was signed on the public link (Phase 16F PR1; the quote analogue of
        downloadCurrentInvoiceSignature). The bytes are the stored server-normalised PNG,
        streamed verbatim: the upload's decoded pixels re-encoded at acceptance (same dimensions
        and pixels; ancillary chunks and trailing bytes dropped), so they are not byte-identical
        to the customer's upload. The workspace accepted.signature_download_path points
        here; consume it verbatim. Inline filename quote-signature-{order_number}-v{n}.png. The
        order is resolved within the session business/store first (an order outside them returns
        404 ORDER_NOT_FOUND; existence is never leaked). No accepted quote version, or no stored
        signature on the latest one, returns 404 (QUOTE_SIGNATURE_NOT_FOUND). A referenced file
        missing from storage returns 500. Allowed when order_status is LAID (read). Portal only -
        never served on the public surface. Not an old-version signature-history endpoint.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Accepted quote signature image streamed
          headers:
            Content-Disposition:
              description: Inline signature filename generated by backend.
              schema:
                type: string
            Content-Length:
              description: Signature file size in bytes.
              schema:
                type: integer
                format: int64
                minimum: 0
          content:
            image/png:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /public/quotes/{token}/pdf:
    servers:
      - url: /api/v1
    get:
      tags:
        - Quotes
      summary: Download the issued quote PDF (public)
      operationId: downloadPublicQuotePdf
      security: []
      description: >
        Streams the issued quote PDF to the customer only while the token is ACTIVE (Phase 16B;
        implemented 16E-C - the stored issued bytes, never regenerated).
        Cost-free. Once the token is CONSUMED (signed) or otherwise dead the endpoint returns 410;
        an unknown token returns 404. The signed PDF is never served here - it is portal only.
      parameters:
        - $ref: '#/components/parameters/Token'
      responses:
        '200':
          description: Issued quote PDF streamed
          headers:
            Content-Disposition:
              description: Inline quote PDF filename generated by backend.
              schema:
                type: string
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '410':
          $ref: '#/components/responses/Gone'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Server-side session cookie used by protected Sales Portal endpoints.
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    Token:
      name: token
      in: path
      required: true
      schema:
        type: string
      description: Opaque secret public quote token (not an order or quote id).
  schemas:
...
    EmptyObjectRequest:
      type: object
      additionalProperties: false
...
    Money:
      type: number
      format: double
      example: 45.00
      description: AUD money amount, returned with two decimal places.
...
    StandardErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/StandardError'
...
    QuoteLineType:
      type: string
      enum:
        - ITEM
        - ADJUSTMENT
...
    QuoteDraftLineInput:
      type: object
      additionalProperties: false
      required:
        - line_type
        - description
        - line_total_ex_gst
        - sort_order
      properties:
        line_type:
          $ref: '#/components/schemas/QuoteLineType'
        description:
          type: string
          minLength: 1
          maxLength: 500
        quantity:
          type: number
          format: double
          nullable: true
        unit_price_ex_gst:
          allOf:
            - $ref: '#/components/schemas/Money'
          nullable: true
        line_total_ex_gst:
          $ref: '#/components/schemas/Money'
        sort_order:
          type: integer
          minimum: 0
    QuoteDraft:
      type: object
      additionalProperties: false
      required:
        - itemised
        - quote_total_ex_gst
        - quote_total_inc_gst
        - below_cost
        - lines
      properties:
        itemised:
          type: boolean
        quote_total_ex_gst:
          $ref: '#/components/schemas/Money'
        quote_total_inc_gst:
          $ref: '#/components/schemas/Money'
        gp_percent:
          type: number
          format: double
          nullable: true
          description: Server-computed GP on the quote total; null when not computable.
        below_cost:
          type: boolean
          description: True when the quote is below cost (save/send/accept are blocked).
        lines:
          type: array
          description: >
            Ordered draft lines. On a non-itemised draft this contains the RETAINED dormant
            itemised rows (empty when none exist; Phase 16D-B PR2A) — they are not part of the
            non-itemised totals and must not be rendered while itemised is false.
          items:
            $ref: '#/components/schemas/QuoteDraftLine'
        updated_at:
          allOf:
            - $ref: '#/components/schemas/Timestamp'
          nullable: true
...
    QuoteWorkspace:
      type: object
      additionalProperties: false
      required:
        - draft
        - current_issued
        - accepted
      properties:
        draft:
          allOf:
            - $ref: '#/components/schemas/QuoteDraft'
          nullable: true
        current_issued:
          allOf:
            - $ref: '#/components/schemas/QuoteIssuedSummary'
          nullable: true
        accepted:
          allOf:
            - $ref: '#/components/schemas/QuoteAcceptedSummary'
          nullable: true
          description: >-
            The order's latest ACCEPTED quote version (Phase 16F PR1), or null when none exists.
            Selected independently of draft and current_issued, so it stays present beside a newer
            draft or a newer issued version.
    QuoteWorkspaceResponseWrapper:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/QuoteWorkspace'
        message:
          type: string
    QuoteDraftUpsertRequest:
      type: object
      additionalProperties: false
      required:
        - itemised
        - lines
      properties:
        itemised:
          type: boolean
        final_total_inc_gst:
          allOf:
            - $ref: '#/components/schemas/Money'
          nullable: true
          description: >
            Desired GST-inclusive final total. Itemised save (optional): if below the line sum the
            server inserts a negative adjustment line to balance; above the line sum returns 422
            (QUOTE_TOTAL_EXCEEDS_LINES). Non-itemised save (required): becomes the quote total
            (ex derived as final / 1.10, HALF_UP 2dp); never validated against retained dormant
            lines (Phase 16D-B PR2A).
        lines:
          type: array
          items:
            $ref: '#/components/schemas/QuoteDraftLineInput'
    QuoteDraftResponseWrapper:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/QuoteDraft'
        message:
          type: string
    QuoteSendResponseWrapper:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/QuoteIssuedSummary'
        message:
          type: string
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    BadGateway:
      description: >
        Bad Gateway. An upstream/provider dependency failed (e.g. the email provider could not
        send the invoice — error.code EMAIL_SEND_FAILED). Phase 13.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    Gone:
      description: >
        Gone. The public quote link is no longer usable - it has expired (QUOTE_LINK_EXPIRED),
        been replaced/superseded (QUOTE_LINK_SUPERSEDED), been cancelled (QUOTE_LINK_CANCELLED),
        or already been signed (QUOTE_LINK_INACTIVE). Phase 16B.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StandardErrorResponse'